+++ src/wallet/CMakeLists.txt
@@ -10,6 +10,7 @@ add_library(bitcoin_wallet STATIC EXCLUDE_FROM_ALL
   crypter.cpp
   db.cpp
   dump.cpp
+  export.cpp
   external_signer_scriptpubkeyman.cpp
   feebumper.cpp
   fees.cpp
